`transferEther` now reverts when transfer fails
... and passes on the revert reason.
The function used to return `false` which isn't too bad because token
transfers can return `false` on failure too and we check the return
value anyway.
Nevertheless, this change makes the `transferEther` function behave more
like Solidity's `transfer`, which also reverts in case of failure and
passes on the revert reason.
Unlike Solidity's transfer, we return `true` in case of success, so we
have no difference in return type between Ether and token transfers.